My_Model implement Soft Delete

+++ application/core/MY_Model.php
@@ -5,8 +5,8 @@ class MY_Model extends CI_Model {
     protected $table_name = '';
     protected $primary_key = 'id';
     protected $database_name = '';
-    
     protected $connection_name = '';
+    protected $soft_delete = false;
     
     //example: $where_override = array('client_id' => $this->override_id);
     //example: $override_column = 'client_id';    
@@ -50,7 +50,15 @@ class MY_Model extends CI_Model {
 */
     }
     public function get($id) {
-        return $this->db->get_where($this->table_name, array($this->primary_key => $id))->row();
+	    if ($this->soft_delete == true)
+        	$this->db->where('delete', 0);
+	    
+	    if($this->where_override)
+			$this->db->where($this->where_override);
+		
+		$this->db->where($this->primary_key, $id);
+	    
+        return $this->db->get($this->table_name)->row();
     }
 
     public function get_all($fields = '', $where = array(), $table = '', $limit = '', $order_by = '', $group_by = '', $join_table = '', $join_where = '', $join_method='left') {
@@ -61,6 +69,9 @@ class MY_Model extends CI_Model {
 
 		if($this->where_override)
 			$this->db->where($this->where_override);
+			
+		if ($this->soft_delete == true)
+        	$this->db->where('delete', 0);
 		
         if (count($where)) {
             $this->db->where($where);
@@ -73,7 +84,7 @@ class MY_Model extends CI_Model {
         if ($join_table != '' && $join_where != '') {
             $this->db->join($join_table, $join_where, $join_method);
         }
-
+		
         if ($limit != '') {
             $this->db->limit($limit);
         }
@@ -105,6 +116,9 @@ class MY_Model extends CI_Model {
 
 		if($this->where_override)
 			$this->db->where($this->where_override);
+		
+		if ($this->soft_delete == true)
+        	$this->db->where('delete', 0);
 			
 // 		$this->db->where(array('client_id' => $this->client_id));
 		$this->db->where(array('last_update >' => $last_update));
@@ -185,22 +199,39 @@ class MY_Model extends CI_Model {
 
     public function delete($id) {
         $this->db->where($this->primary_key, $id);
-
-        return $this->db->delete($this->table_name);
+        
+        if ($this->soft_delete == false)
+            return $this->db->delete($this->table_name);
+        
+        
+        $data['delete'] = 1;
+        $data['status'] = 0;
+        $data['last_update'] = date('Y-m-d H:i:s');
+//         $data['delete_by'] = $this->user_id;
+        
+        return $this->db->update($this->table_name, $data);
     }
     public function delete_array($params) {
-        return $this->db->delete($this->table_name, $params);
+	    if ($this->soft_delete == false)
+	    	return $this->db->delete($this->table_name, $params);
+	    
+    	$params['delete'] = 1;
+    	$params['status'] = 0;
+        $params['last_update'] = date('Y-m-d H:i:s');
+	    
+        return $this->db->update($this->table_name, $params);
     }
     
     public function query($query){
         return $this->db->query($query)->result();
 	}
 	public function query_as_array_auto($query, $arguments = NULL){
-			$data = array();
-//         $this->db->where(array('client_id' => $this->client_id));
+		$data = array();
 
 		if($this->where_override)
 			$this->db->where($this->where_override);
+		if ($this->soft_delete == true)
+        	$this->db->where('delete', 0);
 			
         $query = $this->db->query($query, $arguments);
 		foreach ($query->result_array() as $row)
@@ -216,6 +247,8 @@ class MY_Model extends CI_Model {
         
 		if($this->where_override)
 			$this->db->where($this->where_override);
+		if ($this->soft_delete == true)
+        	$this->db->where('delete', 0);
 			
         $query = $this->db->query($query, $arguments);
 		foreach ($query->result_array() as $row)
@@ -225,29 +258,27 @@ class MY_Model extends CI_Model {
         return $data;
 	}
 	
-	public function replace($data) {
+	       
+    public function replace($data) {
 		$data['last_update'] = date('Y-m-d H:i:s');
-	    //$data['created_from_ip'] = $data['updated_from_ip'] = $this->input->ip_address();
-	// 		$data['client_id'] = $this->client_id;
-	
-		if($this->override_column && $this->override_id)
-		{
-			$data[$this->override_column] = $this->override_id;
-		}
 		
-	    $success = $this->db->replace($this->table_name, $data);
-	    if ($success) {
-	        return $this->db->insert_id();
-	    } else {
-	        return FALSE;
-	    }
+		if($this->override_column && $this->override_id)
+		$data[$this->override_column] = $this->override_id;
+		
+		$success = $this->db->replace($this->table_name, $data);
+		
+		if ($success)
+			return $this->db->insert_id();
+		else 
+			return FALSE;
     }
     
-	public function count_all()
+    public function count_all()
     {
-	    $table = $this->table_name;
-	    $query = "SELECT count(id) as count FROM $table";
-	    $result = $this->query_as_array_auto($query, null);
-	    return $result[0]['count'];
+       $table = $this->table_name;
+       $query = "SELECT count(id) as count FROM $table";
+       $result = $this->query_as_array_auto($query, null);
+       
+       return $result[0]['count'];
     }
 }
